requirements.txt
gpytorch==1.8.1
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
POT==0.9.4
pyro_ppl==1.8.6
rational_activations==0.2.0
scikit_learn==1.1.2
scipy==1.9.0
torch==2.2.1
tqdm==4.66.1
normflows